TS: MatchPresenceFilter
